move API functions to a module
