+++ assets/javascript/modules/api/buildPhotographer.js
@@ -1,5 +1,5 @@
 import getPhotographers from './getPhotographers.js';
-import { Photographer } from './photographer.js';
+import { Photographer } from '../../utils/photographer.js';
 
 async function buildPhotographer() {
     const api = await getPhotographers();
