+++ assets/javascript/modules/api/sumLikes.js
